Добавлена проверка на существование файла при смене аватарки с google ссылки на загруженный аватар.

+++ Yotalab.PlanningPoker.BlazorServerSide/Services/FilesStoraging/AvatarStorage.cs
@@ -48,7 +48,8 @@ namespace Yotalab.PlanningPoker.BlazorServerSide.Services.FilesStoraging
       if (!DefaultAvatarCollection.Contains(avatarDescriptor.RelativeFileName))
       {
         var fullPath = Path.Combine(this.webRootPath, avatarDescriptor.RelativeFileName);
-        File.Delete(fullPath);
+        if (File.Exists(fullPath))
+          File.Delete(fullPath);
       }
     }
